+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.7.13
 ## mm/dd/2021
 
+1. [](#new)
+    * Added support for getting translated collection of Flex Pages using `$collection->withTranslated('de')`
 1. [](#improved)
     * Moved `gregwar/Image` and `gregwar/Cache` in-house to official `getgrav/Image` and `getgrav/Cache` packagist packages. This will help environments with very strict proxy setups that don't allow VCS setup. [#3289](https://github.com/getgrav/grav/issues/3289)
     * Improved XSS Invalid Protocol detection regex [#3298](https://github.com/getgrav/grav/issues/3298)